feat(rpc): add evm namespace and trait bindings
