+++ crates/rpc/rpc-api/src/lib.rs
@@ -7,9 +7,9 @@
 //! - `client`: Enables JSON-RPC client support.
 
 #![doc(
-html_logo_url = "https://raw.githubusercontent.com/paradigmxyz/reth/main/assets/reth-docs.png",
-html_favicon_url = "https://avatars0.githubusercontent.com/u/97369466?s=256",
-issue_tracker_base_url = "https://github.com/paradigmxyz/reth/issues/"
+    html_logo_url = "https://raw.githubusercontent.com/paradigmxyz/reth/main/assets/reth-docs.png",
+    html_favicon_url = "https://avatars0.githubusercontent.com/u/97369466?s=256",
+    issue_tracker_base_url = "https://github.com/paradigmxyz/reth/issues/"
 )]
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
@@ -22,6 +22,7 @@ mod engine;
 mod eth;
 mod eth_filter;
 mod eth_pubsub;
+mod ganache;
 mod mev;
 mod net;
 mod otterscan;
